+++ src/model/Palestra.java
@@ -3,7 +3,7 @@ package model;
 public class Palestra extends Evento {
     private Palestrante palestrante;
 
-    public Palestra(String titulo, String data, int duracaoHoras, double orcamento, Local local, Palestrante palestrante) {
+    public Palestra(String titulo, String data, int duracaoHoras, float orcamento, Local local, Palestrante palestrante) {
         super(titulo, data, duracaoHoras, orcamento, local);
         this.palestrante = palestrante;
     }
